extracted out presiction results into a seperate component
